Create initial implementation of LinkedList
A LinkedList initializes with a Head and Tail reference (default to nil)
and as the list populates, Head will point to the first node in the list
and Tail to the last one.

The method append was implemented in addition to this in order to give
an initial means of populating the list by means of adding a node to the
end of the list.