Add #pop function to delete node at end of list
The node at the end of the list isn't manually deleted, instead the two
references to it are reassigned: @tail is assigned to point to the node
before the last one, and that node's next_node is assigned to point to
nil. Ruby garbage collection should take care of the rest.