BUGFIX for #pop
This method tried to remove a node by setting the tail node to nil, but
failed to account for the edge case of the list being size one, thus
making calls to a nil object. It now properly checks for lists with one
node in it and empties out both head and tail.