Change size from variable tracking to a method
The size instance variable has been replaced with a method of the same
name. This way, instead of manually incrementing/decrementing it as a
variable in methods that add or remove nodes, a method will iterate
through the list and return an integer.

The original idea behind the variable was to keep the size tracking
light-weight in theory, but ultimately puts burden on always manually
updating the variable every time you implement a new feature that adds
or removes a node. In retrospect, it feels silly to not have done it
like this in the first place, but that's what practice is for.